Removed functions from where they don't need to be
We don't need to update the turn in the game-board component. That was a
bad name anyway. Removed this function. Also we didn't need the variable
playerFireTurn so that was removed as well. All it did was call the
updateTurn function.

In the auth service I changed the function name from updateTurn to
getGameId to better represent what the function does.