Anonymous sign in routes to login page.
Anonymous sign in works and navigates to the login page. Running into a
problem where the localhost:4200/login screen does not appear when
logged in this way. The anonymous user is logged in the console. It is
not pushed to the database. @luebelein Will need your help here. This is
some templating code that I don't understand.