Creating a game also creates a chat room.
Created the data model that represents a chat room for a networked game.

chat-rooms is a root collection inside of Cloud Firestore.
Inside of the collection are documents whose id is the userId.
This creates a one to one relationship where a game has one chat room.

This document has a sub collection called messages.
It contains the message that was sent along with the sender's name.

@luebelein What do you think of this data structure?