Code clean up and Chat functionality
Cleaned up some of the code and wrote the function that will send a chat message.

Angular is doing something weird where it is not recognizing the function that I wrote as being defined.

So the function is not being called. Need somebody to take a look at that. It would be called in the gameboard component.html.