Skeleton for create and signIn account functions
Error isn't being thrown when an account that doesn't exist tries to login. I feel it has something to do with promises. The routing works correctly.

Need help with getting the data to post to the database.
Getting a warning in the javascript console that the form submission was canceled because the form isn't connected.

Making good progress.